if dependencies are missing for an adapter, continue without support for that adapter
