+++ dbt/adapters/base/__init__.py
@@ -1,4 +1,4 @@
 from dbt.adapters.base.meta import available
 from dbt.adapters.base.relation import BaseRelation
-from dbt.adapters.base.connections import BaseConnectionManager
+from dbt.adapters.base.connections import BaseConnectionManager, Credentials
 from dbt.adapters.base.impl import BaseAdapter
